# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Zynk API
  description: |
    End-to-end encrypted messaging platform API.
    
    ## Authentication
    All authenticated endpoints require a Bearer token in the `Authorization` header.
    Tokens are obtained via `/auth/login` or `/auth/register`.
    
    ## Rate Limiting
    - Global: 100 requests / 15 minutes per IP
    - Login: 5 attempts / 15 minutes per IP
    - Registration: 3 attempts / 15 minutes per IP
    - Messages: 60 per minute per user
    - File uploads: 20 per hour per user
    
    ## Encryption
    All message content is end-to-end encrypted using the Signal Protocol.
    The server only stores encrypted ciphertext.
  version: 1.0.0
  contact:
    name: Zynk Team
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required: [error]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Device:
      type: object
      properties:
        id:
          type: string
          format: uuid
        device_name:
          type: string
        platform:
          type: string
        last_active_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 64
                  pattern: '^[a-zA-Z0-9_]+$'
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                device_name:
                  type: string
                  default: Web Browser
                device_fingerprint:
                  type: string
                public_key:
                  type: string
              required: [username, password]
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: { type: string, format: uuid }
                  username: { type: string }
                  session_token: { type: string }
                  refresh_token: { type: string }
                  expires_at: { type: integer }
        '409':
          description: Username taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
  /auth/login:
    post:
      tags: [Auth]
      summary: Login to existing account
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
                device_fingerprint: { type: string }
                device_name: { type: string, default: Web Browser }
              required: [username, password]
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: { type: string, format: uuid }
                  session_token: { type: string }
                  refresh_token: { type: string }
                  device_id: { type: string, format: uuid }
                  expires_at: { type: integer }
        '401':
          description: Invalid credentials
        '403':
          description: Max devices reached
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  code: { type: string, enum: [MAX_DEVICES_REACHED] }
                  max_devices: { type: integer }
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
        '429':
          description: Rate limited
  /auth/force-login:
    post:
      tags: [Auth]
      summary: Remove a device and login when at max device limit
      operationId: forceLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
                remove_device_id: { type: string, format: uuid }
                device_fingerprint: { type: string }
                device_name: { type: string }
              required: [username, password, remove_device_id]
      responses:
        '200':
          description: Force login successful
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh session token
      operationId: refreshToken
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_token: { type: string }
                  refresh_token: { type: string }
                  expires_at: { type: integer }
        '401':
          description: Invalid refresh token
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout current session
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Logged out
  /auth/logout-all:
    post:
      tags: [Auth]
      summary: Logout from all devices
      operationId: logoutAll
      security:
        - BearerAuth: []
      responses:
        '204':
          description: All sessions terminated
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user info
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /auth/devices:
    get:
      tags: [Auth]
      summary: List connected devices
      operationId: listDevices
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Device list
          content:
            application/json:
              schema:
                type: object
                properties:
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
  /auth/devices/{deviceId}:
    delete:
      tags: [Auth]
      summary: Remove a device
      operationId: removeDevice
      security:
        - BearerAuth: []
      parameters:
        - name: deviceId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Device removed

  # ======================== Messages ========================
...